+++ src/services/content-org/learning-paths.ts
@@ -8,6 +8,7 @@ import { addContextToContent } from '../contentAggregator.js'
 import {
   contentStatusCompleted,
   contentStatusReset,
+  getAllCompletedByIds,
   getProgressState,
 } from '../contentProgress.js'
 
@@ -220,6 +221,35 @@ export async function fetchLearningPathLessons(
   }
 }
 
+/**
+ * For an array of contentIds, fetch any content progress with state=completed,
+ * including other learning paths and a la carte progress.
+ *
+ * @param contentIds The array of content IDs within the learning path
+ * @param learningPathId The learning path ID
+ * @returns {Promise<Object>} Response object
+ * @returns {Array} result.lessons - Array of all learning path lesson contentIds.
+ * @returns {Array} result.completed_lessons - Array of learning path lesson contentIds that are completed.
+ * @returns {Array} result.lessons_count - Count of learning path lessons.
+ * @returns {Array} result.completed_lessons_count - Count of learning path completed lessons.
+ */
+export async function fetchLearningPathProgressCheckLessons(
+  contentIds: number[],
+  learningPathId: number
+): Promise<object> {
+  let query = await getAllCompletedByIds(contentIds, {
+    id: learningPathId,
+    type: 'learning-path-v2',
+  })
+  let completedContentIds = query.data
+  return {
+    lessons: contentIds,
+    completed_lessons: completedContentIds,
+    lessons_count: contentIds.length,
+    completed_lessons_count: completedContentIds.length,
+  }
+}
+
 interface completeMethodIntroVideo {
   intro_video_response: Object | null,
   active_path_response: ActiveLearningPathResponse
@@ -273,6 +303,7 @@ export async function completeLearningPathIntroVideo(introVideoId: number, learn
     response.learning_path_reset_response = await contentStatusReset(learningPathId, collection)
 
   } else {
+      response.lesson_import_response = {}
     for (const contentId of lessonsToImport) {
       response.lesson_import_response[contentId] = await contentStatusCompleted(contentId, collection)
     }